+++ ovos_workshop/app.py
@@ -25,8 +25,8 @@ from ovos_utils.sound import wait_while_speaking
 
 from ovos_utils.skills import get_non_properties
 from ovos_utils.intents import IntentBuilder, Intent, AdaptIntent
-from ovos_workshop.skills.decorators import *
-from ovos_workshop.skills.decorators.killable import killable_event, \
+from ovos_workshop.decorators import *
+from ovos_workshop.decorators.killable import killable_event, \
     AbortEvent, AbortQuestion
 from ovos_workshop.skills.layers import IntentLayers
