+++ ovos_workshop/decorators/converse.py
@@ -1,0 +1,11 @@
+
+
+def converse_handler():
+    """Decorator for aliasing a method as the converse method"""
+
+    def real_decorator(func):
+        if not hasattr(func, 'converse'):
+            func.converse = True
+        return func
+
+    return real_decorator
