+++ ovos_workshop/skills/decorators/converse.py
@@ -1,11 +1,2 @@
-
-
-def converse_handler():
-    """Decorator for aliasing a method as the converse method"""
-
-    def real_decorator(func):
-        if not hasattr(func, 'converse'):
-            func.converse = True
-        return func
-
-    return real_decorator
+from ovos_workshop.decorators.converse import *
+# backwards compat import
