+++ ovos_workshop/skills/decorators/fallback_handler.py
@@ -1,8 +1,2 @@
-
-def fallback_handler(priority=50):
-    def real_decorator(func):
-        if not hasattr(func, 'fallback_priority'):
-            func.fallback_priority = priority
-        return func
-
-    return real_decorator
+from ovos_workshop.decorators.fallback_handler import *
+# backwards compat import
